+++ src/main/java/hageldave/fftw/ezfftw/samplers/RowMajorArraySampler.java
@@ -5,15 +5,15 @@ import hageldave.fftw.ezfftw.Utils;
 public class RowMajorArraySampler implements RealValuedSampler {
 
 	public final double[] array;
-	private final int[] dimensions;
+	private final long[] dimensions;
 
-	public RowMajorArraySampler(double[] array, int... dimensions) {
+	public RowMajorArraySampler(double[] array, long... dimensions) {
 		this.array = array;
 		this.dimensions = dimensions;
 	}
 
 	@Override
-	public double getValueAt(int... coordinates) {
-		return array[Utils.indexFromCoordinates(coordinates, dimensions)];
+	public double getValueAt(long... coordinates) {
+		return array[(int)Utils.indexFromCoordinates(coordinates, dimensions)];
 	}
 }
